# backend/docs/OPENAPI_SCHEMA.yaml
openapi: 3.0.3
info:
  title: Media Center Arti Marziali API
  description: |
    REST API for the Media Center Arti Marziali platform.

    ## Authentication
    Most endpoints require a Bearer token in the Authorization header.

    ## User Tiers
    - `free` - Basic access with ads
    - `hybrid_light` - No ads, 1080p
    - `hybrid_standard` - Downloads enabled
    - `premium` - 4K, all features
    - `business` - Bulk licenses, API access
  version: 1.0.0
  contact:
    name: API Support
    email: api@example.com
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    VideoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [technique, kata, combat, theory, workout, demo, other]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        style:
          type: string
        tags:
          type: array
          items:
            type: string
        thumbnail_url:
          type: string
        duration:
          type: integer
        view_count:
          type: integer
        tier_required:
          type: string
        created_at:
          type: string
          format: date-time
    VideoListResponse:
      type: object
      properties:
        videos:
          type: array
          items:
            $ref: '#/components/schemas/VideoResponse'
        total:
          type: integer
        skip:
          type: integer
        limit:
          type: integer
    VideoCreateRequest:
      type: object
      required:
        - title
        - category
        - difficulty
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [technique, kata, combat, theory, workout, demo, other]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        style:
          type: string
        tags:
          type: array
          items:
            type: string
        tier_required:
          type: string
          default: free
        is_premium:
          type: boolean
          default: false
        ppv_price:
          type: integer
        instructor_name:
          type: string
...
paths:
...
  /videos:
    get:
      tags: [videos]
      summary: List videos
      operationId: listVideos
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: category
          in: query
          schema:
            type: string
            enum: [technique, kata, combat, theory, workout, demo, other]
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced, expert]
        - name: search
          in: query
          schema:
            type: string
            minLength: 2
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, view_count, title]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: List of videos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoListResponse'

    post:
      tags: [videos]
      summary: Create video (admin)
      operationId: createVideo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoCreateRequest'
      responses:
        '201':
          description: Video created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoResponse'
...
  /videos/{video_id}:
    get:
      tags: [videos]
      summary: Get video details
      operationId: getVideo
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoResponse'
        '404':
          description: Video not found

    put:
      tags: [videos]
      summary: Update video (admin)
      operationId: updateVideo
      security:
        - BearerAuth: []
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video updated

    delete:
      tags: [videos]
      summary: Delete video (admin)
      operationId: deleteVideo
      security:
        - BearerAuth: []
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video deleted
...